requirements-dev.txt
APScheduler==3.6.3
black==21.9b0
cachetools==4.2.2
certifi==2021.10.8
click==8.0.3
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.3
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
python-dateutil==2.8.2
python-decouple==3.5
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.10.23
Sastrawi==1.0.1
six==1.16.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
